chore: rename directory to fit naming convention
